+++ mythtv/libs/libmythtv/infodialog.cpp
@@ -31,24 +31,24 @@ InfoDialog::InfoDialog(ProgramInfo *pginfo, QWidget *parent, const char *name)
     setGeometry(0, 0, screenwidth, screenheight);
     setFixedSize(QSize(screenwidth, screenheight));
 
-    setFont(QFont("Arial", 16 * hmult, QFont::Bold));
+    setFont(QFont("Arial", (int)(16 * hmult), QFont::Bold));
     setCursor(QCursor(Qt::BlankCursor));
 
     QPushButton *ok = new QPushButton("OK", this, "close");
-    ok->setFont(QFont("Arial", 20 * hmult, QFont::Bold));
+    ok->setFont(QFont("Arial", (int)(20 * hmult), QFont::Bold));
 
     QPushButton *cancel = new QPushButton("Cancel", this, "cancel");
-    cancel->setFont(QFont("Arial", 20 * hmult, QFont::Bold));
+    cancel->setFont(QFont("Arial", (int)(20 * hmult), QFont::Bold));
 
     connect(ok, SIGNAL(clicked()), this, SLOT(okPressed()));
     connect(cancel, SIGNAL(clicked()), this, SLOT(reject()));
 
-    QVBoxLayout *vbox = new QVBoxLayout(this, 20 * wmult);
+    QVBoxLayout *vbox = new QVBoxLayout(this, (int)(20 * wmult));
 
-    QGridLayout *grid = new QGridLayout(vbox, 4, 2, 10 * wmult);
+    QGridLayout *grid = new QGridLayout(vbox, 4, 2, (int)(10 * wmult));
     
     QLabel *titlefield = new QLabel(pginfo->title, this);
-    titlefield->setFont(QFont("Arial", 25 * hmult, QFont::Bold));
+    titlefield->setFont(QFont("Arial", (int)(25 * hmult), QFont::Bold));
 
     QLabel *date = getDateLabel(pginfo);
 
@@ -74,7 +74,7 @@ InfoDialog::InfoDialog(ProgramInfo *pginfo, QWidget *parent, const char *name)
 
     QFrame *f = new QFrame(this);
     f->setFrameStyle(QFrame::HLine | QFrame::Plain);
-    f->setLineWidth(4 * hmult);
+    f->setLineWidth((int)(4 * hmult));
     vbox->addWidget(f);    
 
     QVBoxLayout *middleBox = new QVBoxLayout(vbox);
@@ -111,7 +111,7 @@ InfoDialog::InfoDialog(ProgramInfo *pginfo, QWidget *parent, const char *name)
 
     f = new QFrame(this);
     f->setFrameStyle(QFrame::HLine | QFrame::Plain);
-    f->setLineWidth(4 * hmult);
+    f->setLineWidth((int)(4 * hmult));
     vbox->addWidget(f);   
 
     QHBoxLayout *bottomBox = new QHBoxLayout(vbox);
